backend/requirements/observability.txt
# Observability & Monitoring Requirements
# OpenTelemetry, Prometheus, and structured logging

# OpenTelemetry instrumentation
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-django>=0.42b0
opentelemetry-instrumentation-psycopg2>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0

# Prometheus metrics (already in production.txt, but adding specific version)
prometheus-client>=0.18.0

# Enhanced structured logging
python-json-logger>=2.0.7
